Apply RBAC using Pundit

+++ app/controllers/todo_apps_controller.rb
@@ -1,7 +1,7 @@
 class TodoAppsController < ApplicationController
     include NewTodoAppService
-    skip_before_action :verify_authenticity_token, only: [ :new_index, :create, :new_form, :update, :destroy ]
-    before_action :authorize_request, except: [ :index, :create, :show, :update, :destroy, :new_index, :new_form ]
+    skip_before_action :verify_authenticity_token, only: [ :new_index, :create, :new_form, :update, :destroy, :register, :login ]
+    before_action :authorize_request, except: [ :login, :register ]
     def index
         @todos=NewTodoAppService.get_tasks()
         puts @todos
@@ -51,15 +51,11 @@ class TodoAppsController < ApplicationController
     end
 
     def destroy
+        todo_app=TodoApp.find(params[:id])
+        authorize todo_app
         result=NewTodoAppService.delete_task(params[:id])
         if result[:success]
-            format.turbo_stream do
-                render turbo_stream: turbo_stream.remove("todo_#{@task.id}")
-                end
-            # If the request is for JSON (usually for API calls or AJAX)
-            format.json do
-                render json: { message: "Task #{@task.title} deleted successfully!" }, status: 200
-                end
+            render json: { message: result[:message] }, status: result[:status]
         else
             render json: { message: result[:message] }, status: result[:status]
         end
